requirements.txt
numpy
Pillow
opencv-python
keyboard